# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api": {
      "get": {
        "operationId": "AppController_getHello",
        "summary": "あいさつ取得",
        "description": "あいさつを取得する。",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfGetHelloResponse",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/GetHelloResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "summary": "ユーザ登録",
        "description": "ユーザを登録する。",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfUserResponseDto",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "summary": "ログイン",
        "description": "ログインする。",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfLoginResponseDto",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/LoginResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "ユーザ取得",
        "description": "ユーザを取得する。",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfUserResponseDto",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "ユーザ更新",
        "description": "ユーザを更新する。",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfUserResponseDto",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfUserResponseDto",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/spotify/albums": {
      "get": {
        "operationId": "SpotifyController_findAlbums",
        "summary": "楽曲検索",
        "description": "楽曲名からアルバムを検索する。",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "pagination size",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "description": "pagination offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "albumName",
            "required": true,
            "in": "query",
            "description": "query of album name",
            "schema": {
              "example": "千本桜",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfFindAlbumsResponseDto",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/FindAlbumsResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/spotify/artists": {
      "get": {
        "operationId": "SpotifyController_findArtists",
        "summary": "アーティスト検索",
        "description": "アーティスト名からアーティストを検索する。",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "pagination size",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "description": "pagination offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "artistName",
            "required": true,
            "in": "query",
            "description": "query of artist name",
            "schema": {
              "example": "初音ミク",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfFindArtistsResponseDto",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/FindArtistsResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/spotify/artists/{id}/albums": {
      "get": {
        "operationId": "SpotifyController_findArtistAlbums",
        "summary": "アーティストの楽曲検索",
        "description": "アーティストからアルバムを検索する。",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "pagination size",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "description": "pagination offset",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "ApiCommonResponseOfFindAlbumsResponseDto",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/CommonResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "$ref": "#/components/schemas/FindAlbumsResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "info": {
    "title": "api example",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "example",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CommonResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "実行結果",
            "example": true
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "description": "レスポンス送信日時",
            "example": "2024-03-12T07:45:29.583Z"
          },
          "data": {
            "type": "object",
            "description": "データ"
          },
          "error": {
            "type": "string",
            "description": "エラー内容",
            "example": []
          }
        },
        "required": [
          "success",
          "date",
          "data",
          "error"
        ]
      },
      "GetHelloResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "メッセージ",
            "example": "Hello World!"
          }
        },
        "required": [
          "message"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "name",
            "example": "nickname"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "password",
            "example": "password"
          },
          "name": {
            "type": "string",
            "description": "name",
            "example": "nickname"
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "user id",
            "example": "1"
          },
          "access_token": {
            "type": "string",
            "description": "jwt token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOjEsImVtYWlsIjoidGVzdEBleGFtcGxlLmNvbSIsImlhdCI6MTcyNjY0NzQ5MywiZXhwIjoxNzI2NjQ4NjkzfQ.Ac7bLt8f6JoshNX_V0hbbJZj_3LfGcupY3ffFQRKpak"
          }
        },
        "required": [
          "id",
          "access_token"
        ]
      },
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "password",
            "example": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email",
            "example": "test@example.com"
          },
          "password": {
            "type": "string",
            "description": "password",
            "example": "password"
          },
          "name": {
            "type": "string",
            "description": "name",
            "example": "nickname"
          }
        }
      },
      "FindAlbumsResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "pagination size",
            "example": 20
          },
          "limit": {
            "type": "number",
            "description": "pagination size",
            "example": 20
          },
          "offset": {
            "type": "number",
            "description": "pagination offset",
            "example": 0
          },
          "albums": {
            "description": "album list",
            "example": [
              {
                "id": "1",
                "name": "千本桜",
                "release_date": "2021-01-01",
                "detail": "https://example.com",
                "artists": [
                  {
                    "id": "1",
                    "name": "初音ミク"
                  }
                ]
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "albums"
        ]
      },
      "FindArtistsResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "description": "pagination size",
            "example": 20
          },
          "limit": {
            "type": "number",
            "description": "pagination size",
            "example": 20
          },
          "offset": {
            "type": "number",
            "description": "pagination offset",
            "example": 0
          },
          "artists": {
            "description": "artist list",
            "example": [
              {
                "id": "1",
                "name": "初音ミク"
              }
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "artists"
        ]
      }
    }
  }
}